+++ client/src/component/datafetch/__tests__/button.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+
+import Button from '../button';
+
+
+describe('datafetch', () => {
+  describe('Button', () => {
+    test('renders correctly', () => {
+      const tree = renderer.create(<Button />).toJSON();
+      expect(tree).toMatchSnapshot();
+    });
+  });
+});