+++ CHANGELOG.md
@@ -6,3 +6,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 - Main navigation
+- UI data fetching component